+++ src/main/java/com/mangopay/entities/subentities/PayInPaymentDetailsBankWire.java
@@ -8,6 +8,7 @@ import com.mangopay.entities.BankAccount;
 
 import java.lang.reflect.Type;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -38,6 +39,12 @@ public class PayInPaymentDetailsBankWire extends Dto implements PayInPaymentDeta
      */
     @SerializedName("WireReference")
     private String wireReference;
+
+    /**
+     * Transaction details for BankWire Pay-ins retrieval.
+     */
+    @SerializedName("TransactionDetails")
+    private List<PayInTransactionDetailsBankWire> transactionDetails;
 
     public PayInPaymentDetailsBankWire(Money declaredDebitedFunds, Money declaredFees, BankAccount bankAccount, String wireReference) {
         this.declaredDebitedFunds = declaredDebitedFunds;
@@ -79,6 +86,14 @@ public class PayInPaymentDetailsBankWire extends Dto implements PayInPaymentDeta
 
     public void setWireReference(String wireReference) {
         this.wireReference = wireReference;
+    }
+
+    public List<PayInTransactionDetailsBankWire> getTransactionDetails() {
+        return transactionDetails;
+    }
+
+    public void setTransactionDetails(List<PayInTransactionDetailsBankWire> transactionDetails) {
+        this.transactionDetails = transactionDetails;
     }
 
     /**